requirement.txt
#Django
Django==3.2
djangorestframework==3.13.1

#image
Pillow==9.2.0

#
requests==2.28.1
